+++ CHANGES.rst
@@ -6,7 +6,9 @@
 3.0.1 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Document the ``to_json_representation`` variants and add one
+  that guarantees sorted keys. Make the "fast" variant not dependent
+  on second-chance externalization.
 
 
 3.0.0 (2026-05-07)